% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_income_diff.R
\name{compute_income_diff}
\alias{compute_income_diff}
\title{Compute the difference in the income distribution between a given sample and the full AMELIA dataset.}
\usage{
compute_income_diff(
  sample_income,
  amelia_income_dist,
  buckets,
  plot = F,
  show_mean = F,
  show_median = F,
  sample_type = NULL,
  plot_relevance = F,
  return_full = T
)
}
\arguments{
\item{sample_income}{Vector containing the income values of all the people in the sample}

\item{buckets}{Set vector of income buckets}

\item{plot}{Whether plot should be generated}

\item{show_mean}{Whether the (scaled) mean should be shown in the plot as a threshold}

\item{show_median}{Whether the (scaled) median should be shown in the plot as a threshold}

\item{sample_type}{Sampling method used, only used for the title of the plot}

\item{return_full}{Returns full list of deviations if TRUE, returns only squared deviation if FALSE}

\item{Distribution}{of the AMELIA income in buckets, i.e. the output of the get_bucket_dist function for the full AMELIA dataset.}

\item{label_distance}{Distance between labels, only used for plotting}
}
\value{
If (return_full == TRUE)
- List consisting of the difference between the 'true' income in buckets and the sampled income in buckets in three forms:
  - dist_val = Absolute # of people in sample per bucket
  - dist_prop = Proportion of people in sample per bucket
  - difference_table_prop = Difference in table form (used for analysis of sampling methods)
  - difference_table_prop_scaled = Difference table with scaled values (range = 0,1)
  - squared_deviation = Squared deviation of density per bucket in the sample and true distribution
Else
  - squared_deviation  = Squared deviation of density per bucket in the sample and true distribution
}
\description{
This function is a wrapper for all the other function in this file. Note that we use the 'dist_prop'
attribute for the difference as this is a measure that has been adjusted by the sample size. If
we were to use absolute values, we wouldn't get relevant results as the N for AMELIA is too large.

An important step in this function is the computation of the difference between the 'actual' and the sampled
income distribution. Here, we did not simply subtract the relative densities, as this over-emphasizes the relevance of
deviations in the lower income groups, because the density is highest in these groups to begin with. To counteract this
bias, we compute a dampening factor for each bucket, which is simply 1 - density of the specific bucket in either the sample
or the true base population.
}
