% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bucket_dist.R
\name{get_bucket_dist}
\alias{get_bucket_dist}
\title{Get bucketed income distribution of a sample for a provided set of income buckets}
\usage{
get_bucket_dist(x, buckets)
}
\arguments{
\item{x}{income of sample}

\item{buckets}{vector of buckets based on AMELIA dataset}
}
\value{
a list with amount people per income class (table) and  the proportion of people per income class (prop.table)
}
\description{
This function computes the buckets for a given income sample, which is needed to compare it
to the known information about the income distribution of the full dataset.
}
