% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sample.R
\name{importance_sample}
\alias{importance_sample}
\title{Apply importance sampling to a sample to match a given distribution}
\usage{
importance_sample(sample_diff, amelia_buckets, buckets, data, print_cases = T)
}
\arguments{
\item{sample_diff}{List object from compute_income_diff}

\item{amelia_buckets}{Density of each income bucket in full AMELIA dataset, used for scaling the density per bucket}

\item{buckets}{Income buckets}

\item{data}{Sample data to modify}

\item{print_cases}{Whether function should print # of replacements per bucket}
}
\value{
dataframe with added and removed data points
}
\description{
Function to perform under- and oversampling per income buckets.
Since the density for lower income buckets is much higher than for higher income buckets, we
need to put more emphasis on smaller differences in higher income buckets. For this, we
scale the 'true' density per bucket to 1 and apply the same scaling to the sample values. We then
compute the difference between the two.
}
