% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_sample.R
\name{synth_sample}
\alias{synth_sample}
\title{Apply synthetic sampling using kNN to a sample to match a given distribution}
\usage{
synth_sample(
  sample_diff,
  data,
  amelia_buckets,
  buckets,
  k = 4,
  print_cases = T
)
}
\arguments{
\item{buckets}{Income buckets}

\item{k}{# of neighbours to compare to, defaults to 4}

\item{print_cases}{Whether function should print # of replacements per bucket}

\item{List}{object from compute_income_diff}

\item{Sample}{data to modify}

\item{Density}{of each income bucket in full AMELIA dataset, used for scaling the density per bucket}
}
\value{
Dataframe with added data points
}
\description{
Synthetically generate new data points using the kNN function from the FNN package.
Applies the same scaling of relevance to higher values, but does not generate any new
data points if there are less data points in the given cluster than new ones would
be required.
}
